garbage collector

+++ src/Interpreter/ClassDef.java
@@ -7,13 +7,13 @@ import java.util.Map;
 public class ClassDef
 {
     public String name;
-    public Map<String, Integer> attributes;
+    public Map<String, Long> attributes;
     public Map<String, ClassMethod> methods;
 
     public ClassDef(String name)
     {
         this.name = name;
-        this.attributes = new HashMap<String, Integer>();
+        this.attributes = new HashMap<String, Long>();
         this.methods = new HashMap<String, ClassMethod>();
     }
 }
